date-and-time task 1,2 fix
